Moved all files in lib to nicinfo/lip. This fixes loading issues created by name collisions like require 'ip'
