+++ lib/nicinfo.rb
@@ -1,0 +1,1 @@
+require 'nicinfo/nicinfo_main'
