+++ lib/nicinfo/bsfiles/dns.json
