+++ lib/nicinfo/bsfiles/entity.json
