+++ lib/nicinfo/constants.rb
