+++ lib/nicinfo/demo/domain-dnr.json
