+++ lib/nicinfo/demo/domains.json
