+++ lib/nicinfo/demo/entities.json
