+++ lib/nicinfo/demo/entity-dnr.json
