+++ lib/nicinfo/demo/ip.json
