+++ lib/nicinfo/demo/nameservers.json
