+++ lib/nicinfo/demo/ns-simple.json
