+++ lib/nicinfo/demo/ns-very-simple.json
