+++ lib/nicinfo/demo/ns.json
