+++ lib/nicinfo/demo/simple-ip.json
