+++ lib/nicinfo/demo/simple.json
