+++ lib/nicinfo/enum.rb
