+++ lib/nicinfo/nicinfo_main.rb
@@ -17,21 +17,21 @@ require 'optparse'
 require 'net/http'
 require 'net/https'
 require 'uri'
-require 'config'
-require 'constants'
-require 'cache'
-require 'enum'
-require 'common_names'
-require 'bootstrap'
-require 'notices'
-require 'entity'
-require 'ip'
-require 'ns'
-require 'domain'
-require 'autnum'
-require 'error_code'
+require 'nicinfo/config'
+require 'nicinfo/constants'
+require 'nicinfo/cache'
+require 'nicinfo/enum'
+require 'nicinfo/common_names'
+require 'nicinfo/bootstrap'
+require 'nicinfo/notices'
+require 'nicinfo/entity'
+require 'nicinfo/ip'
+require 'nicinfo/ns'
+require 'nicinfo/domain'
+require 'nicinfo/autnum'
+require 'nicinfo/error_code'
 require 'ipaddr'
-require 'data_tree'
+require 'nicinfo/data_tree'
 begin
   require 'json'
 rescue LoadError
