+++ lib/nicinfo/ns.rb
@@ -12,12 +12,12 @@
 # ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR
 # IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
-require 'config'
-require 'nicinfo_logger'
-require 'utils'
-require 'common_json'
-require 'entity'
-require 'data_tree'
+require 'nicinfo/config'
+require 'nicinfo/nicinfo_logger'
+require 'nicinfo/utils'
+require 'nicinfo/common_json'
+require 'nicinfo/entity'
+require 'nicinfo/data_tree'
 
 module NicInfo
